requirements.txt
anthropic
openai
google-genai
ollama
pydantic
